pyproject.toml
[tool.poetry]
name = "stockwhisper"
version = "0.1.0"
description = ""
authors = ["Nacho114 <17376073+Nacho114@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
yfinance = "^0.2.53"
psycopg2 = "^2.9.10"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
